refactor: 수정 내역 → 단위 테스트 반영
- 가격 데이터 타입 최적화
- prepare 로직 수정에 따른 테스트 수정 및 추가
- 상수 객체 설정으로 가독성 완화

*Trade 객체는 아직 Long 타입 유지

Refs: #88

+++ src/main/java/com/sparta/cream/dto/response/PaymentDetailsResponse.java
@@ -1,5 +1,6 @@
 package com.sparta.cream.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.sparta.cream.domain.entity.Payment;
@@ -11,12 +12,12 @@ public class PaymentDetailsResponse {
 	private final Long id;
 	private final String merchantUid;
 	private final String productName;
-	private final Long amount;
+	private final BigDecimal amount;
 	private final String status;
 	private final String method;
 	private final LocalDateTime paidAt;
 
-	public PaymentDetailsResponse(Long id, String merchantUid, String productName, Long amount, String status, String method, LocalDateTime paidAt) {
+	public PaymentDetailsResponse(Long id, String merchantUid, String productName, BigDecimal amount, String status, String method, LocalDateTime paidAt) {
 		this.id = id;
 		this.merchantUid = merchantUid;
 		this.productName = productName;
