+++ src/main/java/com/sparta/cream/dto/response/SettlementDetailsResponse.java
@@ -1,5 +1,6 @@
 package com.sparta.cream.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.sparta.cream.domain.entity.Settlement;
@@ -12,9 +13,9 @@ import lombok.Getter;
 @AllArgsConstructor
 public class SettlementDetailsResponse {
     private final Long id;
-    private final Long feeAmount;
-    private final Long settlementAmount;
-    private final Long totalAmount;
+    private final BigDecimal feeAmount;
+    private final BigDecimal settlementAmount;
+    private final BigDecimal totalAmount;
     private final SettlementStatus status;
     private final LocalDateTime settledAt;
     private final String productName;
