+++ src/main/java/com/sparta/cream/dto/response/YourPaymentListResponse.java
@@ -1,5 +1,6 @@
 package com.sparta.cream.dto.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.sparta.cream.domain.entity.Payment;
@@ -11,11 +12,11 @@ public class YourPaymentListResponse {
 	private final Long id;
 	private final String merchantUid;
 	private final String productName;
-	private final Long amount;
+	private final BigDecimal amount;
 	private final LocalDateTime paidAt;
 	private final String status;
 
-	public YourPaymentListResponse(Long id, String merchantUid, String productName, Long amount, LocalDateTime paidAt,
+	public YourPaymentListResponse(Long id, String merchantUid, String productName, BigDecimal amount, LocalDateTime paidAt,
 		String status) {
 		this.id = id;
 		this.merchantUid = merchantUid;
